server/requirements.txt
fastapi
uvicorn[standard]
pymupdf
python-multipart
google-genai
pydantic
google.generativeai
reportlab
python-docx
